requirements.txt
pip-chill==1.0.3
groq==0.20.0
fastapi==0.115.11
pydantic==2.10.6
python-dotenv==1.0.1
pinecone==6.0.2
pandas==2.2.3
uuid==1.30
uvicorn==0.34.0
python-multipart==0.0.20